Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2569 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qmimap4/src/offlinejob.cpp
@@ -661,7 +661,7 @@ bool qmimap4::SetFlagsOfflineJob::write(OutputStream* pStream) const
 	if (!OfflineJob::write(pStream))
 		return false;
 	
-	unsigned int nSize = listUid_.size();
+	unsigned int nSize = static_cast<unsigned int>(listUid_.size());
 	WRITE(&nSize, sizeof(nSize));
 	WRITE(&listUid_[0], listUid_.size()*sizeof(UidList::value_type));
 	WRITE(&nFlags_, sizeof(nFlags_));
